MVP 2.0 버전 배포

+++ src/main/java/Funssion/Inforum/domain/post/memo/service/MemoService.java
@@ -3,8 +3,9 @@ package Funssion.Inforum.domain.post.memo.service;
 import Funssion.Inforum.common.constant.Sign;
 import Funssion.Inforum.common.constant.DateType;
 import Funssion.Inforum.common.constant.OrderType;
-import Funssion.Inforum.common.exception.ArrayToListException;
-import Funssion.Inforum.common.exception.BadRequestException;
+import Funssion.Inforum.common.exception.etc.ArrayToListException;
+import Funssion.Inforum.common.exception.badrequest.BadRequestException;
+import Funssion.Inforum.common.exception.etc.UnAuthorizedException;
 import Funssion.Inforum.common.utils.SecurityContextUtils;
 import Funssion.Inforum.domain.member.entity.MemberProfileEntity;
 import Funssion.Inforum.domain.mypage.exception.HistoryNotFoundException;
@@ -50,24 +51,23 @@ public class MemoService {
     private final MyRepository myRepository;
     private final S3Repository s3Repository;
 
-    @Transactional(readOnly = true)
-    public List<MemoListDto> getMemosForMainPage(DateType date, OrderType orderBy) {
+    public List<MemoListDto> getMemosForMainPage(DateType date, OrderType orderBy, Long pageNum, Long memoCnt) {
 
         Integer days = DateType.toNumOfDays(date);
 
-        return getMemos(orderBy, days);
+        return getMemos(orderBy, days, pageNum, memoCnt);
     }
 
-    private List<MemoListDto> getMemos(OrderType memoOrderType, Integer days) {
+    private List<MemoListDto> getMemos(OrderType memoOrderType, Integer days, Long pageNum, Long memoCnt) {
         switch (memoOrderType) {
             case NEW -> {
-                return memoRepository.findAllOrderById()
+                return memoRepository.findAllOrderById(pageNum, memoCnt)
                         .stream()
                         .map((MemoListDto::new))
                         .toList();
             }
             case HOT -> {
-                return memoRepository.findAllByDaysOrderByLikes(days)
+                return memoRepository.findAllByDaysOrderByLikes(days, pageNum, memoCnt)
                         .stream()
                         .map(MemoListDto::new)
                         .toList();
@@ -111,7 +111,6 @@ public class MemoService {
                 .toList();
     }
 
-    @Transactional(readOnly = true)
     public MemoDto getMemoBy(Long memoId) {
 
         Memo memo = memoRepository.findById(memoId);
@@ -126,15 +125,16 @@ public class MemoService {
         Long userId = AuthUtils.getUserId(UPDATE);
         ArrayList<String> updatedTags = new ArrayList<>(form.getMemoTags());
         Memo savedMemo = memoRepository.findById(memoId);
+        checkPermission(userId, savedMemo);
+
         updateHistory(form, userId, savedMemo);
         try {
             tagRepository.updateTags(memoId,updatedTags);
         } catch (SQLException e) {
             throw new ArrayToListException("tag update 중 sql.Array를 List로 변환하는 중 오류 발생",e);
         }
-        Memo memo = memoRepository.updateContentInMemo(form, memoId);
 
-        return new MemoDto(memo);
+        return new MemoDto(updateMemo(memoId, form, savedMemo));
     }
 
     private void updateHistory(MemoSaveDto form, Long userId, Memo savedMemo) {
@@ -149,11 +149,23 @@ public class MemoService {
 
     }
 
+    private Memo updateMemo(Long memoId, MemoSaveDto form, Memo savedMemo) {
+        Memo memo;
+        // 실제 메모로 등록된 적이 없는 메모를 등록하려 할 때
+        if (!form.getIsTemporary() && savedMemo.getIsTemporary() && !savedMemo.getIsCreated())
+            memo = memoRepository.updateContentInMemo(form, memoId, Boolean.TRUE);
+        else
+            memo = memoRepository.updateContentInMemo(form, memoId);
+        return memo;
+    }
+
     @Transactional
     public void deleteMemo(Long memoId) {
-
         Long userId = AuthUtils.getUserId(DELETE);
         Memo memo = memoRepository.findById(memoId);
+
+        checkPermission(userId, memo);
+
         try {
             tagRepository.deleteTags(memoId);
         } catch (SQLException e) {
@@ -166,6 +178,12 @@ public class MemoService {
         if (!memo.getIsTemporary())
             myRepository.updateHistory(userId, MEMO, MINUS, memo.getCreatedDate().toLocalDate());
     }
+
+    private static void checkPermission(Long userId, Memo savedMemo) {
+        if (!userId.equals(savedMemo.getAuthorId()))
+            throw new UnAuthorizedException("메모를 업데이트할 권한이 없습니다.");
+    }
+
 
     public ImageDto uploadImageInMemo(Long id, MultipartFile image) {
         Long userId = AuthUtils.getUserId(UPDATE);
@@ -180,7 +198,7 @@ public class MemoService {
                 .imagePath(uploadedURL)
                 .build();
     }
-  
+
     public List<MemoListDto> searchMemosBy(
             String searchString,
             Long userId,
